pyproject.toml
[tool.poetry]
name = "studioblog"
version = "0.1.0"
description = ""
authors = ["Rick Henry <rickhenry@rickhenry.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
wagtail = "^5.1.2"
dj-database-url = "^2.1.0"
whitenoise = "^6.5.0"
webauthn = "^1.11.0"
django-htmx = "^1.16.0"
django-stubs = "^4.2.4"
django-stubs-ext = "^4.2.2"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
pylint = "^3.0.1"
djlint = "^1.34.0"
pre-commit = "^3.4.0"
pyproject-flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true

[tool.django-stubs]
django_settings_module = "studiowiki.settings"

[tool.isort]
skip = [".gitignore", ".dockerignore"]
skip_glob = [".venv/*"]
profile = "black"

[tool.flake8]
max-line-length = 88
